Relocate history components
