+++ rt_dashboard/javascript/app.js
@@ -7,5 +7,4 @@ import './src/components/history'
 import './src/components/queues'
 import './src/components/rt-dashboard'
 import './src/components/workers'
-import './elements/history-chart'
-import './components/history-page'
+import './src/components/history'
