+++ rt_dashboard/javascript/src/components/history/HistoryChart.js
