+++ rt_dashboard/javascript/src/components/history/sample_data.js
